prjoect: Creating adventure game that allows user to pick destinations to get out of a cave
The game involves the user to choose thier destinations through a number of choices.
Each choice the user makes will either get them safely out of the cave or even more lost.
At this point I've created 3 classes, Item, Gold, and Weapon. Items is the superclass and gold is a subclass. Weapons is an extension of Item with a damage attribute. Then Weapon will extend to define some specific weapons
in the game